fix: 모바일 UI 캐시 메시지 버그 수정
- ConfidenceBadge useState 선언 순서 수정 (MISTAKES.md Components #2)
- ChartContent에서 IChartApi 직접 import 제거 및 useChartSync 훅으로 분리 (MISTAKES.md Layer Dependencies #3)
- useChartSync 커스텀 훅 신규 생성 (components/chart/hooks/)
- StockChart, VolumeChart에 onChartReady 콜백 연결
